refactor: break free from cyclical dependencies ss-053
